LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/selection-value-interactions-expected.txt


FAIL value dirty flag behavior after setRangeText on textarea not in body assert_equals: Calling setRangeText should set the value dirty flag expected "somexyzing" but got "set range text"
FAIL value dirty flag behavior after setRangeText on input not in body assert_equals: Calling setRangeText should set the value dirty flag expected "somexyzing" but got "set range text"
PASS value dirty flag behavior after setRangeText on textarea in body 
PASS value dirty flag behavior after setRangeText on input in body 
PASS value dirty flag behavior after setRangeText on textarea in body with parsed default value 
PASS value dirty flag behavior after setRangeText on input in body with parsed default value 
PASS value dirty flag behavior after setRangeText on focused textarea 
PASS value dirty flag behavior after setRangeText on focused input 
PASS value dirty flag behavior after setRangeText on focused then blurred textarea 
PASS value dirty flag behavior after setRangeText on focused then blurred input 
PASS selection is always collapsed to the end after setting values on input 
FAIL selection is always collapsed to the end after setting values on textarea assert_equals: element.selectionStart should be value.length expected 3 but got 0
